+++ src/Table/Livewire/TreeModels.php
@@ -5,6 +5,7 @@ namespace Thinktomorrow\Chief\Table\Livewire;
 use Thinktomorrow\Chief\Managers\Exceptions\MissingResourceRegistration;
 use Thinktomorrow\Chief\Resource\TreeResource;
 use Thinktomorrow\Chief\Shared\Concerns\Nestable\NestableTree;
+use Thinktomorrow\Vine\DefaultNode;
 
 class TreeModels
 {
@@ -16,7 +17,9 @@ class TreeModels
     public function compose(TreeResource $resource, array $ids, int $offset, int $limit, string $keyName = 'id'): array
     {
         // Get the entire tree structure but only the ids to reduce memory load
-        $treeModelIds = NestableTree::fromIterable($resource->getTreeModelIds());
+        $treeModelIds = NestableTree::fromIterable($resource->getTreeModelIds(), function ($model) use ($keyName) {
+            return new DefaultNode($model, new NestableTree, $keyName, 'parent_id');
+        });
 
         // Reduce the tree to only the models that are expected by the current filtering / sorting.
         $treeIds = collect(
@@ -34,7 +37,7 @@ class TreeModels
 
         // Add indent to models based on the tree depth
         $models = $allModels
-            ->filter(fn ($model) => in_array($model->id, $modelIds))
+            ->filter(fn ($model) => in_array($model->{$keyName}, $modelIds))
             ->map(function ($model) use ($treeModelIds, $keyName) {
                 $model->indent = $treeModelIds->find(fn ($node) => $node->getNodeId() == $model->{$keyName})->getNodeDepth();
 
@@ -42,7 +45,7 @@ class TreeModels
             })->values();
 
         $ancestors = $allModels
-            ->filter(fn ($model) => in_array($model->id, $ancestorIds))
+            ->filter(fn ($model) => in_array($model->{$keyName}, $ancestorIds))
             ->map(function ($model) use ($ancestorTreeIds, $keyName) {
                 $model->indent = $ancestorTreeIds->first(fn ($node) => $node->getNodeId() == $model->{$keyName})->getNodeDepth();